scope addonTreeCache to app instance
It's not really correct to share them across broccoli builders, because the trees get torn down in between.

It also turned out there was no point in importing the cache out of addon.js, because it caches different content anyway (it always caches recursive treeFor, not our own flat treeFor).