fix stage2 externals resolution
It is somewhat common for apps and/or addons to try to import things from other addons that are not listed in `dependencies` (or for apps, `devDependencies`). As stage1 compiles each package, it will flag these things as `externals`.

Then at stage2, we double-check to see if any of the externals refer to other ember addons that are present in the build. If they do, we remove them from the externals list, and allow them to statically link up.

This PR fixes that linkage so it works for any path imported from an addon (like `@ember-decorators/object/computed`) not just the package's own top level (like `@ember-decorators/object`).