importSync macro
This replaces the previous `import { require } from '@embroider/core'` macro that was internal-only.

`importSync` is intended to be a public macro.

Ember itself really already needs synchronous dynamic imports to exists, so we're not really imposing any extra burden on ourselves to make this a public macro. And making it public helps with #150.