change smooshing order
When combining multiple copies of the same addon, ember-cli gives priority to ones closer to the app in the dependency graph. That is, for each addon it builds the dependencies first, then overwrites the dependencies with the addon itself.

Our addon discovery returns each addon before its dependencies, so we were going in the opposite priority.