tolerate sloppy addon overwriting
It's not great for addons to overwrite their own trees, but it's common enough that we probably need to tolerate it.

A common example is an addon that accidentally emits an (unprocessed) CSS file from treeForAddon, only for it to be overwritten by a (processed) CSS file from the actual style trees.