Move AppBuilder’s writes to the shared macro config to before user land add-ons are setup.
I’m not 100% sure this is the appropriate fix, but
from what I can tell, once we traverse user addons the config is not longer safely updatable.

Addons, such as ember-cli-fastboot, may interact with the preprocessor registry in a way the forces the babel config to be accessed. Accessing the babel config, will also access the macro config, as babel most likely must be aware of the macro config to work correctly, the act of accessing the macro config “freezes” it, as subsequent  writes may be lost.

We should likely discuss additional approaches, and see if I’m missing something important.