compatibility with new ember-cli-htmlbars
The inline-precompile plugin has moved into ember-cli-htmlbars, so we need to expand our detector.

We always detect and replace the inline template precompile because we need to use our own template compiler instead (to get component and helper dependency resolution, and because ours supports a serialization strategy that works inside webpack's thread-loader).