minor cleanup
I noticed this code is now unnecessary because the latest draft RFC doesn't require addons to namespace their public assets.

It was considered in earlier drafts, but URLs are not always free to change, so v2 packages keep the same powers that v1 packages have. However, we are still a little better at being strict because we error if there are collisions.