refactoring to support eager recursive macro evaluation
This makes it possible for all the macros to be eagerly evaluated within Evaluator. That allows us to use higher and earlier entrypoints from within the babel visitor. For example, the `each` macro can now apply its changes at enter instead of exit, which means it won't be impacted by preset-env.

This also fixes the original problem we were trying to fix here, which was allowing optional chaining on the getConfig macro. We can now evaluate the optional chaining ourselves and it will simply down all the way into the macros.