expose macros babel install
If an addon wants to emit code containing macros into its parent package, it needs to ensure that the parent package will be processed by the macros babel plugin.

This is only an issue in classic builds, because of the way they scope babel configs. Under embroider, babel plugins aren't scoped per package because they're all supposed to be safe, correct language extensions.

The first use case for this is ember-classic-decorator, which is going to emit `isDevelopingThisPackage` macros.