adding support to addons that use treeFor() to suppress various trees
In general, we can't run custom `treeFor()` hooks because they expect to recurse into all their child addons and have visibility into the output of all their child addons, which we forbid.

But the most common use case for `treeFor()` is quite simple, and we *can* opportunistically handle that case. It's the case where treeFor is returning either falsy or the unchanged output of `_super.treeFor()`.

This letse us detect those cases and suppress trees that are supposed to get suppressed.