Ensure dependencySatisfies works for prereleases.
Based on https://github.com/npm/node-semver#prerelease-tags:

> If a version has a prerelease tag (for example, 1.2.3-alpha.3) then it
> will only be allowed to satisfy comparator sets if at least one
> comparator with the same [major, minor, patch] tuple also has a
> prerelease tag.

That means that if an addon uses:

```js
dependencySatisfies('ember-source', '^3.0.0')
```

When the project actually has `3.2.0-beta.1` installed
`semver.satisifies` will _by default_ return `false`.

For our purposes however, we need to ensure that is allowed.  Mostly
because the "end user" (the host project, or the package that included
the package that uses `dependencySatisfies`) has explicitly asked for a
prerelease version.