Renaming addon package to util
The new `@emroider/addon` package is confusingly named. I'm changing it to `@embroider/util` because it's really not just for addons.

I picked "addons" originally because I was thinking of several other things I want to put in it specifically for addons, but that stuff doesn't even exist yet and what does exist is definitely good for apps too.