fix caching bug
The default implementation of `cacheKeyForTree` checks whether there are any customized methods that back the given tree name.

We use that same system for our own v2Trees, because it interacts correctly with addon's own customnized cache key implementations.

But we have been using the cacheKey for the `addon` tree only, and that is not correct, because if the addon tree is not customized, ember-cli will give us a key for it, even though the addon may have other customized trees that should cause the v2Tree to really not be cacheable.