adding @babel/plugin-transform-runtime by default
This makes all apps smaller by sharing the babel helper functions within a library, instead of inlining them many times.

This plugin needs to be sensitive to which files are ES modules (because those `import` the helpers) and which are CJS (because those `require` the helpers). Unfortunately for compatibility we do allow both to coexist within our build, so there is a possibility some weirder use cases may break, which we will need to keep an eye on.