allowUnsafeDynamicComponents flag
This can be helpful while testing, because instead of stopping your build at the first potentially-missing component, you can get a complete build and see how much of your app is working in the browser.

It is not recommended to be on in production, because it makes it easy to accidentally ship runtime errors.