Avoid running CI jobs on both push and pull_request events.
When pushing changes to a branch, CI would run **double** the jobs (one
set of CI jobs for the branch push event, and another set of jobs for
the `pull_request` event).

This change tweaks the CI hooks to only run for pull requests
**targeting** `master` and branches named `master` (also allow `v*`
style release branches, just in case we wanted them in the future).