Remove ember-cli-htmlbars dependencies in @embroider/compat
We still need `ember-cli-htmlbars@3` for tests (specifically to ensure
that inline precompilation works still in that mode), but we definitely
do not need to ship this dependency to our consumers.

The only import we had for `ember-cli-htmlbars` itself was specifically
for types, which we define internally (in
`types/ember-cli-htmlbars/index.d.ts`).