Respect fingerprinting options during stage 3
We currently were fingerprinting all assets when in production
mode which would break testing as testem.js would become fingerprinted
and thus cause problems loading. This threads through the fingerprinting
options from the app and then determines if it should fingerprint or
not from that. This makes the behavior inline with classic builds.