move peerDep check into activeAddonChildren
I'm suggesting this change on top of https://github.com/embroider-build/embroider/pull/728 because I'd rather keep this behavior local to child addon discovery and not make it be something that's part of the Package's API.

This should still provide the same performance benefit, because even though we will now examine the peerDeps before filtering them out, that work is guarded by the PackageCache and won't actually be repeated unnecessarily. The main thing we're trying to avoid is building many copies of each peerDep per consuming addon.