Ensure extracted CSS comes before app/vendor CSS
CSS that webpack loads from JS imports currently gets added after the vendor and app CSS. This makes it difficult for addons or apps to customize this CSS, e.g. by overriding CSS variables or rules, because they cannot rely on document order for such overrides, so have to play tricks with specificity or the dreaded `!important`.

This commit moves such CSS to just before the first in-`index.html` stylesheet, ensuring that it comes before both the vendor and app CSS.

This also splits out the CSS handling from the JS handling a little more to make the render() method easier to follow, mainly by making it clear that the fastboot replacement stuff only applies to JS and not CSS.

Fixes #801