Fix babel plugin cache eviction
Today, babel’s caches (via babel-loader today) expires caches only if
the options passed to it change, unfortunately if you upgrade a plugin
without changing the configuration options and the file contents remains
the same, the cache will not be evicted. This can lead to spurious
caching problems if the associated plugins behavior has changed.

To address this defect we now include (when available) the
package.json#version of the babel plugins used in our own no-op babel
plugin’s configuration. This new no-op plugin is then inserted at the
end of the configuration in question.