Undefined does not serialize under broccoli-babel-transpiler
When broccoli-babel-transpiler tries to determine in a plugin is
serializable it checks for certain types:
https://github.com/babel/broccoli-babel-transpiler/blob/416732dd5d57b9f29dad7afd7f7cae76f0ac1606/lib/parallel-api.js#L45-L61.

If the value is undefined it is not serializable and thus not
parallelizable. Since we are just checking for truthyness of the appPackageRoot
value we can instead use an empty string. This also satisfies the babel
type which is string | undefined.