Deflate AdjustImportsOptions
The babel-loader cache key hashes babel options which has the
AdjustImportsOptions used multiple times and is very large.

This PR moves this into a file ignoring caching for now since everything
it moves is already memoized and only thing this affects is
in-repo-addons and isDevelopingAddon true which is currently an issue.

This doesn't address relocatedFiles which is very large because it is
outside the spot that is memoized.