Tighten job timeouts
Speaking with the team, we would love to reach a 10 minute Commit -> CI
feedback loop goal. (Assuming the GH Actions queue depth isn't too deep)

We have spoken with  GH to increase our concurrency, but in-order to
reach our goal each of our jobs cannot exceed 5->8minutes.

This is a first step in locking that in.