[Bugfix release] EmberENV.EXTEND_PROTOTYPES is not always present
Newer versions of ember and embroider are not currently compatible, as EmberENV.EXTEND_PROTOTYPES is no longer present.
Luckily, the offending code in embroider is the ember-inflector compact-adapter and as of ember-inflector@4.0.0 is no longer needed.

To support both pre and post ember-inflector@4.0.0 we conditionally execute the compatibility code when safe & required