backing out ember real-modules mode
 - it breaks many addons that aren't ready for the global `Ember` to be missing during vendor.js
  - it exacerabtes the problem of ember's internally-packaged deps (like route_recognizer) fighting with other copies

We discussed making this a flag instead, but since then the plan is that Ember 4.0 itself will make ember-auto-import 2.0 mandatory, which smoothes the path to distributing ember-source as a v2 addon, in which case that should become the "flag" for doing static analysis of ember.

Instead, all that was required was a compat adapter for `@glimmer/tracking`, since it's magic bullshit used to be handled by the modules-api-polyfill and now it needs to be handled differently.