beginning to standardize all template compilation as JS
That is, instead of having different code paths for standalone hbs files vs templates embedded inside Javascript, we do a small transportation to the standalone hbs files to make them into JS first, and then share all the template compilation machinery.

This is also setting us up to take advantage of lexical scope in non-strict-mode which landed in ember-source 3.28.4.