Fix unsafe reuse of broccoli trees in OneShot
The OneShot transform consumes a broccoli tree a single time, detaching us permanently from future updates to that tree, which avoids needing to revalidate all the third-party v1-to-v2 addon conversion during rebuilds.

Since OneShot fully consumes a broccoli tree in a separate broccoli pipeline, it's not safe to consume that same tree again in a different pipeline.

But an addon can manipulate cacheKeyForTree such that two distinct copies of an addon emit the same broccoli tree instance as their v2Tree, which we send into OneShot. This fails in a very byzantine way.

It's the source of the problems people have been having with test-waiters (https://github.com/embroider-build/embroider/issues/1056) because that addon aggressively tries to deduplicate itself.

The fix is to not rerun OneShot if we encounter an identical tree a second time.