exclude inner node modules
When we encounter a v2 addon that has v1 addon as dependencies, we need to move it even tough it doesn't need to be rewritten. Our code for doing that excludes node_modules, but we should also exclude inner nested node_modules.

This comes up when testing a v2 addon that nests its test-app inside itself using symlinked node_modules. Without this you encounter a symlink cycle that's hard to diagnose and only strikes on some OS / disk combinations.