Make addon-shim a non-ember-addon
This changes `@embroider/addon-shim` from being an ember addon to being a plain NPM library.

All v2 addons are required to depend on `@embroider/addon-shim` in order to interoperate with the rest of ember-cli.

A v2 addon that has any v1 addons as dependencies needs to get copied into our temporary work area so that it can resolve those v1 addons *after* they've been compiled to v2. Whereas a v2 addon with only v2 dependencies can stay inert in node_modules and impose nearly zero built-time cost.

This meant that having addon-shim as a v1 addon was de-optimizing all v2 addons. And we can't do addon-shim as a v2 addon because it explicitly exists to hook into the classic Addon instance machinery.

Along with this change, I had to also adjust the resolving rules so that v2 addons that get to stay in the real node_modules see the correct ember virtual peer deps.