Adjust lts scenarios
Adding an Ember 3.28 LTS to the test matrix. We used to test is as "release" but now that has moved on.

And in order to keep the suite the same size, I'm dropping the scenario for 3.20. We still test 3.16, our oldest supported LTS, which should give us plenty of coverage. Things that work in 3.16 and 3.24 are very unlikey to be broken in 3.20.