implement relativeToAssetDir + trim superfluous `assets/` paths
relative references to scripts and styles parsed from HTML entrypoints have
their paths normalised using HTMLEntrypoint.relativeToApp to remove the
leading rootURL set by the app. this results in all resources mapping
directly to their output location on disk. However, when we eventually need
to write the replacement script and link tags for these resources into the
output HTML we need to trust that the publicAssetURL config value provided
by the app is the source of truth for where the assets will be served from
with no intermediate `assets/` directory - so we strip any leading `assets/`
if it exists