Move dummy app support down into packageCache
We have pretty good support for classic Dummy apps in Embroider builds, but our PackageCache infrastructure is also used in classic builds (by @embroider/macros and by ember-auto-import) and there it ends up failing to understand dummy apps consistently.

Critically, a dummy app needs to be represented by a different Package from the actual NPM package it lives inside, because the dummy app has access to devDependencies and the addon does not.

This PR is refactoring the dummy app support down a layer directly into the PackageCache.