Port file assertions to qunit
This makes our internal file assertions utility work in qunit as well as jest.

We have a mix of both due to different generations of development. New stuff is qunit for consistency with out-of-the-box ember runtime tests.

This also uses the assertions to improve the the v2-addon-dev scenario, as well as slightly expanding the coverage there to cover a template transforms, which previously didn't have any end-to-end coverage.