fix a rebuild crash in dummy apps on windows
Rebuilds could crash on windows when working with a v1 addon's dummy app.

We already defend against the weird circularity of dummy apps, but the check didn't work on windows.

This has test coverage that will land as part of #1246, because that includes a refactor of the test suite that reveals this problem.