add code-equality-assertions and upgrade jest
Adopting code-equality-assertions so that we don't get burned by non-semantic changes in output formats.

In order to adopt it, I had to upgrade jest because jest brings its own resolver (sigh) that didn't understand newer node `exports` in package.json.

I had to use resolutions to deduplicate graceful-fs because jest pulled that in and it breaks support for `process.chdir` when there's more than one copy.