defend against infinite loop on broken babel config
It's not legal to have an empty array as a plugin in a babel config -- that definitely makes babel blow up. But in that case, our serialization code runs first and can get trapped in an infinite loop on the bad plugin.

This prevents the loop so you can get to the real babel crash instead.