allow v2 addons to use app tree fallback resolution
We are careful to not allow most of the bad backward-compatible kinds of package resolution inside v2 addons.

But there's an exception that we need to allow and don't: due to app tree merging, files from a v2 addon can end up needing to resolve back to that v2 addon from the app, even when the v2 addon is not properly resolvable from the app.

This is already handled for v1 addons through a process of attempting to resolve from the app, from the original addon, and then from the global list of all known addons. This change allows v2 addons to also use that process, but only when resolving their own name.