test-setup: use caret version modifier for Embroider dependencies
When all monorepo packages were released in-sync, this ensured test-util is refering always to the latest. But with the release process decoupled now, test-util would effectively pull in an older pinned version of Embroider. This adds the caret modifier to allow pulling the latest version (on the same major).

Fixes #1268.