Refactor self-resolution
Before we turned self-resolutions into absolute filenames, now we use the newer rehome() option in the module-resolver. This unblocks the next step in refactoring app-tree- merging into the module-resolver, because when a request like these ones fails we need to do the fallback and would prefer to see the non-absolute specifier still.