clarify which package rules apply inside vs outside a component
Previously we were keying all component rules off the actual on-disk file for that component, even when those rules were being consume at *invocation sites* of the component rather than inside the component itself. This meant that we could not apply those rules if we move component resolution out of the AST transform stage.

But it's actually OK to key the exterior rules off the component's globally-addressable (or at least engine-scoped) name, since the point of the rules is to control classic global resolution.