upgrade typescript
All our stuff actually typechecked -- the first try! -- under typescript 5.04 as well. But rollup-plugin-ts doesn't support it, and when I tried to use a different typescript only in the package that needs that (`@embroider/router`) pnpm did the wrong thing with peerDeps so it didn't actually work. So 4.9.0 is newer-enough for now.