Removing flaky test
Mitigates #1465.

These tests are fundamentally not a good way to test this. It's going to be very hard to make them reliable.

It would probably be better to use [rollup.watch](https://rollupjs.org/javascript-api/#rollup-watch) directly because it emits events we can use, instead of trying to parse stderr and use arbitrary delays to guess when changes have propagated.