changing expectRewrittenFiles so it can handle more complex input paths
The problem with the previous attempt is that it assumes all deps live directly under node_modules in the input. But we have plenty of tests that do more complicated things than that, like using nested npm dependencies or in-repo addons.

This changes the API so that we always talk about locations in the input scenarios and then map them reliably through RewrittenPackageCache to assert about their output locations.