[wip] working to reform or remove AppDiffer
This is a side-quest for #1435. The engines-related tests fail there because we still have app-tree-merging support for dumping addon files into engines, which requires us to manage two copies of each engine. Since our module-resolver can now handle all the app-tree-merging virtually, we can drop that whole part of AppDiffer (and maybe *all* of AppDiffer) to solve the problem.

And reforming AppDiffer makes this a good time to remove the Stage concept..

I'm doing this on a separate PR because this is going to take a step backard in terms of passing tests before it moves forward again. I expect the first commit breaks some fastboot-specific features in apps (we already handle those same features in addons via the module-resolver).