Eagerly virtualize emberVirtualPackages
... and expand the list of virtual pacakges to includes some that are private-but-used-API.

Before this, we only provided these as fallbacks when normal resolving failed, and this mostly worked. But that was when we were also rewriting the entire node_modules graph on disk. Now that we don't do that, there are many more opportunities to accidentally resolve copies of packages that aren't really what you expected.