less aggressive realpathSync
The realpath checks were added because in some environments node_modules is symlinked and we need to account for that. But they also mean that if any directory that appears in the index isn't present it can cause a crash, and that can indeed happen when output from a previous build is still present and hasn't been updated yet.

This should hopefully achieve both goals by only resolving the top-level workingDir, and that only after proving that it's present.

Not doing realpath on every individual inner directory is OK because we control those.