move eager modules before amdModules
When using importSyncImplementation:cjs, the amdModules are evaluated lazily, after the eager modules.

When using importSyncImplementation:eager (which is needed for vite and any true ES module environment), everything is going to get evaluated in the order it is written here. So we should write the eager modules before the lazy ones.