Fix rewritten package cache encapsulation
Fixes https://github.com/embroider-build/embroider/issues/1514

A v2 addon with a v1 addon dependency was not resolving its dependencies correctly, in two places.
 - in virtual-content.ts where we emit implicit-modules
 - in compat-app-builder.ts when doing partitionEngines

The fix to both is adjusting the meaning of RewrittenPackageCache so
that it always automatically provides packages that see moved
dependencies, even when a given package *itself* has not moved (because
it's a v2 addon).