Use rollup's `addWatchFile` API to mark dependencies
This solves a bunch of watch-mode (not) rebuilding issues; as a
general rule everytime we probe or read from the file system we
should at least consider whether such an annotation is needed.

Unfortunately this package doesn't appear to be tested currently,
but I have applied the same patch locally and it appears to be
working fine.